ultimos cambios : Agregado del seed altenativo, controladores probados, agregado de excepciones , etc

+++ Contract/Dto/Request/ClientRequest.cs
@@ -6,7 +6,7 @@
         public string LastName { get; set; }
         public string Email { get; set; }
         public int Cuit { get; set; }
-        public int Dni { get; set; }
+        public long Dni { get; set; }
         public DateTime Nacimiento { get; set; }
         public int Phone { get; set; }
     }
